==> 02-descriptive.Rmd <==
# Descriptives

Load the required packages.

```{r packages}
#| results: 'hide'
#| cache: FALSE

library(knitr)
library(scales)
library(gtsummary)
library(flextable)
library(lme4)
library(broom.mixed)
library(kableExtra)
library(here)
library(tidyverse)
library(lubridate)
```

We then load the previously cleaned data table.

```{r data-load}
data_path <- here("Data", "cleaned_data.rds")
if (file.exists(data_path)) {
  d <- read_rds(file = data_path)
} else {
  stop(str_glue("{data_path} doesn't exist, run `01-clean.Rmd` to create it."))
}

# Make wave a nicely labelled factor
d <- d %>%
  mutate(Wave = factor(wid, levels = 1:3, labels = paste0("Wave ", 1:3)))

# Rename game to fit titles in plots
d <- d %>% 
  mutate(Game = if_else(Game == "Gran Turismo Sport", "GT Sport", Game))
```

## Demographics after exclusions

```{r demographics-table}
d %>%
  filter(wid == 1) %>%
  distinct(pid, Game, Age, Experience, Gender) %>%
  select(-pid) %>%
  tbl_summary(by = Game, missing_text = "Missing") %>%
  add_overall() %>%
  bold_labels() %>%
  italicize_levels() %>%
  as_kable_extra(caption = "Sample demographics") %>% 
  kable_styling(full_width = FALSE, font_size = 12)
```

## Survey descriptives

These are for people after exclusions.

### Response rate & retention

```{r response-rate-table}
# Get data on invite dates and Ns
invites <- read_csv(here("Data", "invites.csv")) %>%
  rename(Game = game) %>%
  mutate(
    Game = str_replace(Game, "Animal Crossing: New Horizons", "AC:NH"),
    Game = str_replace(Game, "Gran Turismo Sport", "GT Sport")
    )
# Create a table where wave 0 are number of invites,
# then calculate response rate / retention at each wave.
# This assumes there are no new participants at wave 3
# (people who didn't participate in wave 2 showed up at wave 3).
tmp <- bind_rows(
  select(invites, -date),
  d %>% filter(Responded) %>% count(Game, wid)
)
tmp_total <- tmp %>% 
  group_by(wid) %>% 
  summarise(n = sum(n)) %>% 
  mutate(Game = "Total")
tmp <- bind_rows(tmp, tmp_total) %>% 
  arrange(Game, wid) %>%
  group_by(Game) %>%
  mutate(
    R_rate = percent(n / lag(n), .1),
    n = comma(n)
  ) %>%
  pivot_wider(names_from = wid, values_from = c(n, R_rate)) %>%
  mutate(
    Invites = n_0,
    `Wave 1` = str_glue("{n_1} ({R_rate_1})"),
    `Wave 2` = str_glue("{n_2} ({R_rate_2})"),
    `Wave 3` = str_glue("{n_3} ({R_rate_3})")
  ) %>%
  select(Game, Invites:`Wave 3`) %>%
  mutate(across(everything(), ~ str_replace(., "NA", "0"))) %>% 
  ungroup()
tmp %>%   
  kbl(caption = "Number of people (response/retention rate) participating at each wave.") %>% 
  kable_styling(full_width = FALSE, font_size = 12)
# Also save this table for MS
tmp %>% 
  flextable() %>% 
  save_as_docx(path = "Figures/Retention-rates.docx")
```

### Missingness

Here we report on the % missing for each variable

```{r}
install.packages("naniar")
library(naniar)
d %>% 
  select(Game, pid, Wave, Affect, `Life satisfaction`, Hours) %>% 
  complete(nesting(Game, pid), Wave) %>% 
  group_by(Game, Wave) %>% 
  select(-pid) %>% 
  miss_var_summary() %>% 
  select(-n_miss) %>% 
  mutate(pct_miss = percent(pct_miss/100, .1)) %>% 
  pivot_wider(
    names_from = c(Wave, variable), 
    values_from = pct_miss, 
    names_glue = "{variable} ({Wave})"
    ) %>% 
  kbl(caption = "Percent missing at each wave for key variables") %>% 
  kable_styling(full_width = FALSE, font_size = 12)
```


### Differences

We then looked at differences between people who dropped out vs who did not.

```{r}
tmp <- d %>% 
  replace_na(list(Responded = FALSE)) %>%
  group_by(Game, pid) %>% 
  summarise(
    across(
      c(Affect, `Life satisfaction`, 
        Intrinsic, Extrinsic, 
        Age, Experience, Hours),
      mean, na.rm = TRUE
    ),
    dropped = factor(
      sum(Responded) != 3, 
      levels = c(FALSE, TRUE), 
      labels = c("No", "Yes")
    )
  )
tmp <- ungroup(tmp)
tmp <- tmp %>% 
  pivot_longer(Affect:Hours)
tmp %>%   
  ggplot(aes(dropped, value)) +
  stat_summary() +
  stat_summary(fun = mean, geom = "line", group = 1) +
  facet_grid(name~Game, scales = "free_y", margins = "Game")
out <- tmp %>% 
  group_by(name) %>% 
  summarise(
    fit = list(
      lmer(value ~ dropped + (1 + dropped | Game), data = cur_data())
    )
  )
out %>% 
  mutate(
    out = map(fit, ~tidy(.x, "fixed"))
  ) %>% 
  select(-fit) %>% 
  unnest(out) %>% 
  filter(term != "(Intercept)") %>% 
  mutate(across(where(is.numeric), ~number(., .01))) %>% 
  transmute(
    Variable = name,
    Difference = str_glue("{estimate} ({std.error})")
  ) %>% 
  flextable(cwidth = 2) %>% 
  save_as_docx(path = "Figures/Between-dropout-table.docx")
```


### Response dates

Only for actual responses (not rows where survey date was filled to be able to join telemetry)

```{r response-date-histograms}
#| fig.height: 8
#| fig.cap: Histograms of response dates.

d %>%
  # Take only actually responded-to waves
  filter(Responded) %>%
  mutate(Date = as_date(StartDate)) %>%
  count(Game, Wave, Date) %>%
  ggplot(
    aes(Date, n, fill = Wave)
  ) +
  geom_col() +
  scale_y_continuous(
    "Responses",
    breaks = pretty_breaks(),
    expand = expansion(c(0, .1)),
  ) +
  scale_x_date(
    "Date",
    date_breaks = "7 day", date_labels = "%b\n%d", date_minor_breaks = "1 day"
  ) +
  facet_wrap("Game", scales = "free_y", ncol = 1)
```

Response times

```{r response-time-histograms}
#| fig.height: 8
#| fig.cap: Histograms of response times (in UTC).

d %>%
  filter(Responded) %>%
  mutate(Hour = hour(StartDate)) %>%
  count(Game, Wave, Hour) %>%
  ggplot(aes(Hour, y = n, fill = Wave)) +
  scale_y_continuous(
    "Responses",
    breaks = pretty_breaks(),
    expand = expansion(c(0, .1)),
  ) +
  scale_x_continuous(
    breaks = seq(0, 21, by = 3),
    expand = expansion(c(0.01))
  ) +
  geom_col() +
  facet_wrap("Game", scales = "free", ncol = 2) +
  theme(legend.position = "bottom")
```

#### Durations between waves

Participants could respond with variable delays due to variation in email schedules and late responding. So we also check the actual intervals between completing waves. Very small values are possible because a participant could have e.g. completed both waves 2 and 3 in succession after receiving wave 3 invitation. Note that negative values were also possible for this reason but they were excluded before. (This figure is restricted to 5-30 day intervals to display the bulk of the data.)

```{r intervals-between-waves-histogram}
#| fig.height: 8
#| fig.cap: >
#|    Histograms of intervals between participants completing the survey waves (in days).

# Table
d %>%
  select(Wave, Game, interval) %>%
  # group_by(Wave) %>%
  filter(Wave != "Wave 1") %>%
  summarise(
    Value = quantile(
      interval,
      probs = c(0, .10, .25, .5, .75, .90, 1),
      na.rm = T
    ) %>%
      round(3),
    Quantile = percent(c(0, .10, .25, .5, .75, .90, 1))
  ) %>%
  pivot_wider(names_from = Quantile, values_from = Value) %>%
  kbl(caption = "Interval duration percentiles preceding waves 2 and 3.") %>% 
  kable_styling(full_width = FALSE, font_size = 12)

# Figure
d %>%
  filter(Wave != "Wave 1") %>%
  filter(between(interval, 5, 30)) %>% 
  mutate(Wave = fct_drop(Wave)) %>%
  ggplot(aes(interval)) +
  geom_vline(xintercept = 14, size = .2) +
  geom_histogram(binwidth = 1, col = "white") +
  scale_y_continuous(
    "Count",
    expand = expansion(c(0, .1))
  ) +
  scale_x_continuous(
    "Days between responding",
    breaks = pretty_breaks()
  ) +
  facet_grid(Game ~ Wave, scales = "free_y")
```

## System information

```{r}
sessionInfo()
```
